requirements.txt
streamlit
google-generativeai>=0.8.6
google-cloud-translate
python-dotenv
pandas
requests
